// src/config/navBarConfig.ts
import {
	LinkPreset,
	type NavBarConfig,
	type NavBarLink,
	type NavBarSearchConfig,
	NavBarSearchMethod,
} from "../types/config";
import { siteConfig } from "./siteConfig";

// 根据页面开关动态生成导航栏配置
const getDynamicNavBarConfig = (): NavBarConfig => {
	// 基础导航栏链接
	const links: (NavBarLink | LinkPreset)[] = [
		// 主页
		LinkPreset.Home,

		// 归档
		LinkPreset.Archive,
	];

	// 根据配置决定是否添加友链，在siteConfig关闭pages.friends时导航栏不显示友链
	if (siteConfig.pages.friends) {
		links.push(LinkPreset.Friends);
	}

	// 根据配置决定是否添加留言板，在siteConfig关闭pages.guestbook时导航栏不显示留言板
	if (siteConfig.pages.guestbook) {
		links.push(LinkPreset.Guestbook);
	}

	// 我的及其子菜单
	links.push({
		name: "我的",
		url: "/my/",
		icon: "material-symbols:person",
		children: [
			// 根据配置决定是否添加相册，在siteConfig关闭pages.gallery时导航栏不显示相册
			...(siteConfig.pages.gallery ? [LinkPreset.Gallery] : []),

			// 根据配置决定是否添加番组计划，在siteConfig关闭pages.bangumi时导航栏不显示番组计划
			...(siteConfig.pages.bangumi ? [LinkPreset.Bangumi] : []),
		],
	});

	// 企鹅群链接
	links.push({
		name: "企鹅群",
		url: "https://qm.qq.com/q/wrmF4FI9pu",
		external: true,
		icon: "simple-icons:qq",
	});

	// 关于及其子菜单
	links.push({
		name: "关于",
		url: "/content/",
		icon: "material-symbols:info",
		children: [
			// 根据配置决定是否添加赞助，在siteConfig关闭pages.sponsor时导航栏不显示赞助
			...(siteConfig.pages.sponsor ? [LinkPreset.Sponsor] : []),

			// 关于页面
			LinkPreset.About,
		],
	});

	// 自定义导航栏链接,并且支持多级菜单
	links.push({
		name: "链接",
		url: "/links/",
		icon: "material-symbols:link",

		// 子菜单

		children: [
			{
				name: "GitHub",
				url: "https://github.com/fqzlr",
				external: true,
				icon: "fa7-brands:github",
			},
			{
				name: "Vuekooの主页",
				url: "https://home.fqzlr.com",
				external: true,
				icon: "fa7-solid:house-chimney",
		    },
			{
				name: "Vuekooの图床",
				url: "https://tu.fqzlr.com",
				external: true,
				icon: "fa7-solid:image",
			},
			{
				name: "Vuekooの邮局",
				url: "https://yj.fqzlr.edu.kg",
				external: true,
				icon: "fa7-solid:envelope",
			},

			{
				name: "Vuekooの笔记",
				url: "https://bj.fqzlr.com",
				external: true,
				icon: "material-symbols:book-4-rounded",
			},
			{
				name: "Vuekooの统计",
				url: "https://umami.fqzlr.com/share/kHCJG2ZUL1r6q5Js",
				external: true,
				icon: "material-symbols:123",
			},
			{
				name: "Vuekooの状态",
				url: "https://kuma.fqzlr.com/status/1",
				external: true,
				icon: "fa7-solid:chart-line",
			},
		],
	});

	// 仅返回链接，其它导航搜索相关配置在模块顶层常量中独立导出
	return { links } as NavBarConfig;
};

// 导航搜索配置
export const navBarSearchConfig: NavBarSearchConfig = {
	method: NavBarSearchMethod.PageFind,
};

export const navBarConfig: NavBarConfig = getDynamicNavBarConfig();
